pyproject.toml
[project]
name = "modelmint-back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dependency-injector>=4.46.0",
    "httpx>=0.28.1",
    "huggingface-hub>=0.32.0",
    "langchain-openai>=0.3.18",
    "mcp[cli]>=1.9.1",
    "supabase>=2.15.1",
]
